LPS-175383 Better naming and move to the correct package
